Fix sending daily submitted forms to be strictly typed and no more dynamic
