+++ backend/src/StamAcasa.Common/DTO/AddFormModel.cs
@@ -1,10 +1,11 @@
 ﻿using System;
+using StamAcasa.Common.Models;
 
 namespace StamAcasa.Common.DTO
 {
     public class FormInfo
     {
-        public string Content { get; set; }
+        public UserForm SubmitedForm { get; set; }
         public string FormTypeId { get; set; }
         public DateTime Timestamp { get; set; }
         public int UserId { get; set; }
