+++ backend/src/StamAcasa.Common/Mapper/FormMappingProfile.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Newtonsoft.Json;
 using StamAcasa.Common.DTO;
 using StamAcasa.Common.Models;
 
@@ -9,7 +10,10 @@ namespace StamAcasa.Common.Mapper
         public FormInfoMappingProfile()
         {
             CreateMap<Form, FormInfo>()
-                .ForMember(m => m.UserInfo, f => f.MapFrom(src => src.User));
+                .ForMember(m => m.UserInfo, f => f.MapFrom(src => src.User))
+                .ForMember(m => m.SubmitedForm,
+                    f => f.MapFrom(src => JsonConvert.DeserializeObject<UserForm>(src.Content)));
+
         }
     }
 }
